requirements.txt
pandas>=2.0
openpyxl>=3.1
pyyaml>=6.0
fpdf2>=2.7
